Move tests to their respective components
